ARM ld: Initial ARM cross-linker
This has a great chance of being very very broken, but it is sufficient to build
simple linux binaries, linked with this linker, and run them on raspbian.

Anything not directly useful to the kernel (or testing on raspbian) is missing.
There are probably unimplemented but important relocations.  TLS support is
entirely absent (and will assert if you try to use it).