ARM Stub out an initial aboot
As part of getting the kernel boot strapped, we're prototyping a
per-target effective fake bootloader 'aboot' (ARM boot, but pronounced
aboot). The goal here is that this will be an interim thing that maps
the kernel, boot archive, and then sets up to jump into that world with
all the VAs that we'd expect to be set up, set up.

This fleshes out the Makefile system based on the old ARM gate and gives
us a foundation for getting started on other pieces.